refactor: remove desktop and web history and settings modules
