+++ packages/hoppscotch-common/src/lib/sync/environments/index.ts
@@ -120,15 +120,21 @@ async function loadGlobalEnvironments() {
     const globalEnv = res.right.me.globalEnvironments
 
     if (globalEnv) {
-      const globalEnvVariableEntries = JSON.parse(globalEnv.variables)
+      const parsed = JSON.parse(globalEnv.variables)
 
-      const result = entityReference(GlobalEnvironment).safeParse(
-        globalEnvVariableEntries
-      )
+      // Parse ladder: wrapper → wrap-bare-array → empty wrapper.
+      let result = entityReference(GlobalEnvironment).safeParse(parsed)
+
+      if (!result.success && Array.isArray(parsed)) {
+        result = entityReference(GlobalEnvironment).safeParse({
+          v: 2,
+          variables: parsed,
+        })
+      }
 
       runDispatchWithOutSyncing(() => {
         setGlobalEnvVariables(
-          result.success ? result.data : globalEnvVariableEntries
+          result.success ? result.data : { v: 2, variables: [] }
         )
         setGlobalEnvID(globalEnv.id)
       })
@@ -182,7 +188,22 @@ function setupUserEnvironmentUpdatedSubscription() {
       // handle the case for global environments
       if (isGlobal) {
         runDispatchWithOutSyncing(() => {
-          setGlobalEnvVariables(JSON.parse(variables))
+          const parsed = JSON.parse(variables)
+
+          // Mirror the load path: try as-is first (so v0 bare arrays still
+          // migrate via verzod), only wrap a bare array as a v2 envelope
+          // when the as-is parse fails (the broken-intermediate-state case).
+          let result = entityReference(GlobalEnvironment).safeParse(parsed)
+          if (!result.success && Array.isArray(parsed)) {
+            result = entityReference(GlobalEnvironment).safeParse({
+              v: 2,
+              variables: parsed,
+            })
+          }
+
+          setGlobalEnvVariables(
+            result.success ? result.data : { v: 2, variables: [] }
+          )
         })
       } else {
         // handle the case for normal environments
