+++ packages/hoppscotch-selfhost-web/src/api/queries/GetSMTPStatus.graphql
@@ -1,0 +1,3 @@
+query GetSMTPStatus {
+  isSMTPEnabled
+}
