--- packages/hoppscotch-selfhost-web/src/platform/collections/desktop/import.ts
@@ -1,1 +1,0 @@
-export { importToPersonalWorkspace } from "../web/import"
