+++ packages/hoppscotch-selfhost-web/tsconfig.json
@@ -21,7 +21,6 @@
       "@app/composables/*": ["./src/composables/*"],
       "@app/helpers/*": ["./src/helpers/*"],
       "@app/api/*": ["./src/api/*"],
-      "@app/lib/*": ["./src/lib/*"],
       "@app/kernel/*": ["./src/kernel/*"]
     },
     "types": ["vite/client", "unplugin-icons/types/vue"]
